Rolled back CLI chess to a former commit, those files are now in here.
For some reason, the refactoring I had done from commits
after 35d5485 (the last good commit was 10 days ago)
had a severe affect on the runtime. Went gung-ho on the encapsulation
and I realized it resulted in a lot more method calls that necessary.

Being that GUI chess is a 'handle' for CLI chess, it was important
that I have CLI chess running quickly -- meaning as soon as a request
for a move is made, the results come back relatively immediately.