Early stages of undo, works for CLI -- just needs more work for GUI.
If readInput() reads a string of "undo",
the undo will happen on the backend.
I just need to reflect the changes for the GUI pieces.

The same idea will be done with readInput() for saving a game.

If readInput() reads a string of "save",
I will save the current state of the game.
(this can be useful for debugging, as well as the final product).

There is a method in Game.java called startFromFilePlayByPlay(),
that reads an input file of this form:

a1 a2
a7 a6

and so on...

I may make the first four lines of the input file as follows:
<user-defined title>
<start time>
<end time>
<winner>
<then moves begin here, line by line, move for move>

and then the "saved" games would be .txt files,
loaded up into startFromFilePlayByPlay().
(this already works in CLI, you hit ENTER to see the next
move played from the file. ENTER is recognized as "")

The other alternative is to serialize the Game object,
store that in a file -- and create a method to
"replay" the game step by step from the move list.
(this may be better)